requirements.txt
pyusb==1.2.1
PySide2==5.15.2.1